@@ -20,27 +20,27 @@
     "require" : {
         "php" : ">=7.1",
         "mouf/magic-query" : "^1.2.3",
-        "mouf/schema-analyzer": "^1.1.1",
-        "doctrine/dbal": "~2.6",
+        "mouf/schema-analyzer": "^1.1.3",
+        "doctrine/dbal": "^2.6.1",
         "psr/log": "~1.0",
         "doctrine/inflector": "~1.0",
         "beberlei/porpaginas": "~1.0",
         "mouf/classname-mapper": "~1.0",
         "doctrine/cache": "^1.6",
-        "greenlion/php-sql-parser": "^4.1",
-        "phlib/logger": "^2.1",
+        "greenlion/php-sql-parser": "^4.1.2",
+        "phlib/logger": "^2.1.1",
         "symfony/console": "^2.0 || ^3.0",
         "mouf/utils.log.psr.multi-logger": "^1.0",
         "symfony/filesystem": "~2.7 || ~3.0",
-        "ramsey/uuid": "^3.6"
+        "ramsey/uuid": "^3.7"
     },
     "require-dev" : {
-        "phpunit/phpunit": "^5.5",
-        "satooshi/php-coveralls": "^1.0",
+        "phpunit/phpunit": "^5.7.21",
+        "satooshi/php-coveralls": "^1.0.1",
         "wa72/simplelogger" : "^1.0",
-        "friendsofphp/php-cs-fixer": "^2.1",
-        "symfony/process": "^3.0",
-        "couscous/couscous": "^1.6",
+        "friendsofphp/php-cs-fixer": "^2.5",
+        "symfony/process": "^3.3.6",
+        "couscous/couscous": "^1.6.1",
         "thecodingmachine/dbal-fluid-schema-builder": "^1.2.1"
     },
     "suggest" : {
